Allow multiple children to be specified for a node inline
Originally we already allowed omitting the wrapping array for children
when there is only one child, e.g. H[:foo, "bar"]. This is now
generalized to allow any number of children to be defined without having
to wrap them in an array. This sometimes just looks better, especially
regarding indentation.